Extract accounting journal entry view model
